feat(gui): move VTFDecoder to utility folder
