+++ src/gui/utility/VTFDecoder.cpp
