+++ src/gui/utility/VTFDecoder.h
